Move category/execution/ethereum/rlp/decode_error.* to category/core/rlp/decode_error.*
