+++ category/execution/ethereum/core/test/test_int_rlp.cpp
@@ -15,11 +15,11 @@
 
 #include <category/core/byte_string.hpp>
 #include <category/core/int.hpp>
+#include <category/core/rlp/decode_error.hpp>
 #include <category/execution/ethereum/core/rlp/address_rlp.hpp>
 #include <category/execution/ethereum/core/rlp/bytes_rlp.hpp>
 #include <category/execution/ethereum/core/rlp/int_rlp.hpp>
 #include <category/execution/ethereum/rlp/decode.hpp>
-#include <category/execution/ethereum/rlp/decode_error.hpp>
 #include <category/execution/ethereum/rlp/encode2.hpp>
 
 #include <evmc/evmc.hpp>
