+++ category/execution/ethereum/db/util.cpp
@@ -21,6 +21,7 @@
 #include <category/core/likely.h>
 #include <category/core/log.hpp>
 #include <category/core/result.hpp>
+#include <category/core/rlp/decode_error.hpp>
 #include <category/core/runtime/unaligned.hpp>
 #include <category/execution/ethereum/core/account.hpp>
 #include <category/execution/ethereum/core/block.hpp>
@@ -35,7 +36,6 @@
 #include <category/execution/ethereum/core/transaction.hpp>
 #include <category/execution/ethereum/db/util.hpp>
 #include <category/execution/ethereum/rlp/decode.hpp>
-#include <category/execution/ethereum/rlp/decode_error.hpp>
 #include <category/execution/ethereum/rlp/encode2.hpp>
 #include <category/mpt/compute.hpp>
 #include <category/mpt/db.hpp>
