+++ category/execution/ethereum/rlp/decode.hpp
@@ -21,7 +21,7 @@
 #include <category/core/likely.h>
 #include <category/core/result.hpp>
 #include <category/core/rlp/config.hpp>
-#include <category/execution/ethereum/rlp/decode_error.hpp>
+#include <category/core/rlp/decode_error.hpp>
 
 #include <boost/outcome/try.hpp>
